@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,11 +27,11 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  get: ^4.3.8
-  pie_chart: ^3.1.1
-  http: 0.12.2
-  intl: 0.15.8
+  cupertino_icons: ^1.0.4
+  get: ^4.6.1
+  pie_chart: ^5.1.0
+  http: ^0.13.4
+  intl: ^0.17.0
 
 dev_dependencies:
   flutter_test:
